requirements.txt
cosdata-sdk
streamlit
sentence-transformers
requests
python-dotenv
pypdf
pydantic
pydantic-settings
pandas
plotly
